Updated LoreAnimation.java for multiple items.
I have edited LoreAnimations.java so that it can work for any item ID. This is simple to implement. The format remains the same as in the apple.txt, but the name is now changed to ID.txt. I have also moved the inventory list scan from the onEnable method into the onOpenInventory and onCloseInventory methods in order for it to work for new players that join after reload (which shouldn't be used). This should be working, but it is untested. Make a copy of your old cope and then test this code. If there's any issues shout at me in that thread.